created a better component for movies
