+++ src/app/app.module.ts
@@ -10,6 +10,7 @@ import { FooterComponent } from './footer/footer.component';
 import { NgxMaskModule, IConfig } from 'ngx-mask';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { PaginationComponent } from './pagination/pagination.component';
+import { MovieCardComponent } from './movie-card/movie-card.component';
 
 
 export const options: Partial<null | IConfig> | (() => Partial<IConfig>) = null;
@@ -24,6 +25,7 @@ export const options: Partial<null | IConfig> | (() => Partial<IConfig>) = null;
     FooterComponent,
     PageNotFoundComponent,
     PaginationComponent,
+    MovieCardComponent,
 
   ],
   imports: [
