--- src/app/movie-details/movie-details.component.html
@@ -1,42 +1,0 @@
-<div *ngIf="movie && movie.poster_path && movie.overview">
-  <div class="card">
-    <div class="card-image">
-      <img
-        [src]="'https://image.tmdb.org/t/p/w500' + movie.poster_path"
-        alt="{{ movie.title }}"
-      />
-
-      <div class="bookmark-container">
-        <a class="bookmark-btn">
-          <i class="material-icons book-add-icons">bookmark_border</i>
-        </a>
-      </div>
-
-      <div class="fab-container">
-        <a class="btn-floating halfway-fab-custom waves-effect waves-light red">
-          <i class="material-icons">favorite_border</i>
-        </a>
-      </div>
-
-    </div>
-
-    <div class="card-content general-text black">
-      <p class="card-title truncate-title movie-title" title="{{ movie.title }}">
-        {{ movie.title }}
-      </p>
-      <p class="release-date mt-5">
-        <i class="material-icons">date_range</i>
-        <span>{{ formatBrazilianDate(movie.release_date) }}</span>
-      </p>
-      <p class="original-language">
-        <i class="material-icons">language</i>
-        <span>{{ movie.original_language.toUpperCase() }}</span>
-      </p>
-
-      <p class="star-vote-average">
-        <i class="material-icons">star</i>
-        <span>{{ formatStarAverageValue(movie.vote_average) }}</span>
-      </p>
-    </div>
-  </div>
-</div>
