--- src/app/movie-details/movie-details.component.scss
@@ -1,111 +1,0 @@
-.card-title {
-  font-weight: bolder;
-  font-size: 1.2rem;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.7);
-  background-color: rgba(0, 0, 0, 0.5);
-  color: white;
-  text-align: center;
-  max-height: 2rem;
-  width: 100%;
-  display: block;
-  padding: 0.5rem 0;
-  text-overflow: ellipsis;
-}
-
-.truncate-title {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  transition: max-width 0.3s ease-in-out;
-
-  &:hover {
-    max-width: 100%;
-    overflow: visible;
-    white-space: normal;
-  }
-}
-
-.card-content {
-  overflow: auto;
-
-  font-size: 0.8rem;
-  p {
-    font-weight: bold;
-    padding: 3px;
-  }
-  span {
-    font-weight: 400;
-    color: whitesmoke;
-  }
-}
-
-.fab-container {
-  position: absolute;
-  bottom: 10px;
-  right: 10px;
-  display: flex;
-  gap: 10px;
-  transition: background-color 0.3s;
-
-  .halfway-fab-custom {
-    margin-bottom: 0 !important;
-  }
-
-  .btn-floating i {
-    &:hover {
-      background-color: white;
-      color: red;
-    }
-  }
-}
-
-.star-vote-average,
-.original-language,
-.release-date {
-  display: flex;
-  align-items: center;
-  text-align: center;
-  gap: 0.3rem;
-  color: white;
-}
-
-.star-vote-average {
-  i {
-    color: #f5c518;
-  }
-}
-
-.card-image {
-  position: relative;
-  object-fit: cover;
-}
-
-.bookmark-container {
-  position: absolute;
-  top: 0;
-  left: 0;
-  z-index: 3;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.7);
-
-  .bookmark-btn {
-    display: inline-block;
-    color: lightgray;
-    position: relative;
-    overflow: hidden;
-    z-index: 1;
-    width: 40px;
-    height: 40px;
-    line-height: 40px;
-    transition: background-color 0.3s;
-    cursor: pointer;
-    vertical-align: middle;
-
-    i {
-      font-size: 3rem;
-    }
-
-    &:hover {
-      color: gray;
-    }
-  }
-}
