+++ src/app/movie-list/movie-list.component.html
@@ -1,7 +1,7 @@
 <div class="">
   <div class="row">
     <div class="col s10 m3 center" *ngFor="let movie of movies">
-      <app-movie-details [movie]="movie"></app-movie-details>
+      <app-movie-card [movie]="movie"></app-movie-card>
     </div>
   </div>
